% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectationMaximization.R
\name{EM_ZIBB}
\alias{EM_ZIBB}
\title{The integrative function for expectation maximization}
\usage{
EM_ZIBB(
  X_sub,
  N_sub,
  alpha = 1,
  beta = 1,
  delta_gc = 0.5,
  iterations = 10,
  stop_diff = 1e-05,
  gc = "Normal",
  save_path = NULL,
  verbose = T
)
}
\arguments{
\item{X_sub}{alternate allele count matrix.}

\item{N_sub}{total variant count matrix.}

\item{alpha}{parameter of beta prior, estimated from M-step.}

\item{beta}{parameter of beta prior, estimated from M-step.}

\item{delta_gc}{parameter representing non-zero distribution component, estimated from M-step.}

\item{iterations}{number of EM iterations.}

\item{stop_diff}{threshold of stopping the EM process if log likelihood differentce is smaller than the threshold.}

\item{gc}{name of background cluster, used for saving the file.}

\item{save_path}{path to save the fitted parameters. Default setting will store in "zibb_fit" folder under current working directory.}

\item{verbose}{whether to print out the detailed time and debugging information in every 1/10 iterations.}
}
\value{
A list of estimated parameters:
\itemize{
\item delta_gc_list - list of parameters representing non-zero distribution component in each iteration,  estimated from EM process.
\item alpha_list - list of parameters of beta prior in each iteration, estimated from EM process all iterations.
\item beta_list - list of parameters of beta prior in each iteration, estimated from EM process all iterations.
\item LL_list - list of log likelihood in each iteration.
\item delta_ij_hat - matrix representing missing variable Zij
}
}
\description{
The integrative function for expectation maximization
}
\examples{
param_fit = EM_ZIBB(X_sub,N_sub,alpha=1,beta=1,delta_gc=0.5,iterations=100,stop_diff=1e-05,save_path=NULL,gc="normal",verbose=T)
}
