% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectationMaximization.R
\name{Estep}
\alias{Estep}
\title{Expectation Step (E-step) to calculate the expected value of missing variable Zij.}
\usage{
Estep(X_sub, N_sub, alpha, beta, delta_gc)
}
\arguments{
\item{X_sub}{alternate allele count matrix.}

\item{N_sub}{total variant count matrix.}

\item{alpha}{parameter of beta prior, estimated from M-step.}

\item{beta}{parameter of beta prior, estimated from M-step.}

\item{delta_gc}{parameter representing non-zero distribution component, estimated from M-step.}
}
\value{
delta_ij_hat, a matrix representing the expected value of each cell and variant's missing identity Zij.
}
\description{
Expectation Step (E-step) to calculate the expected value of missing variable Zij.
}
\examples{
delta_ij_hat = Estep(X_sub,N_sub,alpha,beta,delta_gc)
}
