% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterVariants.R
\name{FilterVariants}
\alias{FilterVariants}
\title{Filter out variants that are purely noise}
\usage{
FilterVariants(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Filter out variants that are purely noise
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
