% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectationMaximization.R
\name{loglik}
\alias{loglik}
\title{Total log likelihood of the whole model and all data}
\usage{
loglik(X_sub, N_sub, delta_gc, alpha, beta, delta_ij_hat)
}
\arguments{
\item{X_sub}{alternate allele count matrix.}

\item{N_sub}{total variant count matrix.}

\item{delta_gc}{parameter representing non-zero distribution component, estimated from M-step.}

\item{alpha}{parameter of beta prior, estimated from M-step.}

\item{beta}{parameter of beta prior, estimated from M-step.}

\item{delta_ij_hat}{a matrix representing the expected value of each cell and variant's missing identity Zij.}
}
\value{
LL, a numeric value representing total log likelihood
}
\description{
Total log likelihood of the whole model and all data
}
\examples{
new_LL = loglik(X_sub,N_sub,delta_gc,alpha,beta,delta_ij_hat)
}
