% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectationMaximization.R
\name{Mstep}
\alias{Mstep}
\title{Maximization Step in the EM, estimate each parameter based on expectation of Zij from E-step.}
\usage{
Mstep(
  X_sub,
  N_sub,
  delta_ij_hat,
  alpha_old,
  beta_old,
  opt_method = c("optim", "MoM")
)
}
\arguments{
\item{X_sub}{alternate allele count matrix.}

\item{N_sub}{total variant count matrix.}

\item{alpha_old}{parameter of beta prior, from last iteration.}

\item{beta_old}{parameter of beta prior, from last iteration.}

\item{opt_method}{parameter of optimization method for alpha and beta. optim: optimize with R's optim pacakge.}
}
\value{
a list of parameters estimated from the M-step:
\itemize{
\item delta_gc_new - parameter representing non-zero distribution component, estimated from M-step in each iteration.
\item alpha_new - parameter of beta prior, estimated from M-step in each iteration.
\item beta_new - parameter of beta prior, estimated from M-step in each iteration.
}
}
\description{
Maximization Step in the EM, estimate each parameter based on expectation of Zij from E-step.
}
\examples{
params = Mstep(X_sub,N_sub,delta_ij_hat,alpha,beta,opt_method=c("optim"))
}
