% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectationMaximization.R
\name{negative_loglik_ab_related}
\alias{negative_loglik_ab_related}
\title{Negative log likelihood related with alpha and beta}
\usage{
negative_loglik_ab_related(parm, data)
}
\arguments{
\item{parm}{a list containing the rest parameters in the ZIBB model.}

\item{data}{alist containing all data in the ZIBB model.}
}
\value{
NLL, a numeric value representing negative log likelihood related with alpha and beta
}
\description{
Negative log likelihood related with alpha and beta
}
\examples{
NLL=negative_loglik_ab_related(parm,data)
}
