% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagnosticPlots.R
\name{ZIBB_fit_diagnostic_plot}
\alias{ZIBB_fit_diagnostic_plot}
\title{Create diagnostic plot of the ZIBB EM parameter inference process}
\usage{
ZIBB_fit_diagnostic_plot(
  X_sub,
  N_sub,
  delta_gc = 0.5,
  alpha = 1,
  beta = 1,
  delta_ij_hat = 0.5,
  gc = "normal",
  delta_gc_list,
  LL_list,
  alpha_list,
  beta_list,
  save_path = NULL,
  width = 10,
  height = 10
)
}
\arguments{
\item{X_sub}{alternate allele count matrix.}

\item{N_sub}{total variant count matrix.}

\item{delta_gc}{initial value of delta_gc}

\item{alpha}{initial value of alpha}

\item{beta}{initial value of beta}

\item{delta_ij_hat}{matrix representing missing variable Zij}

\item{gc}{name of background cluster, used for saving the file.}

\item{delta_gc_list}{list of parameters representing non-zero distribution component in each iteration,  estimated from EM process.}

\item{LL_list}{list of log likelihood in each iteration.}

\item{alpha_list}{list of parameters of beta prior in each iteration, estimated from EM process all iterations.}

\item{beta_list}{list of parameters of beta prior in each iteration, estimated from EM process all iterations.}

\item{save_path}{path to save the diagnostic plots. Default setting will store in "zibb_fit" folder under current working directory.}

\item{width}{width of saved pdf, unit: inches}

\item{height}{height of saved pdf, unit: inches}
}
\value{
A diagnostic PDF.
}
\description{
Create diagnostic plot of the ZIBB EM parameter inference process
}
\examples{
ZIBB_fit_diagnostic_plot(X_sub,N_sub,delta_gc=0.5,alpha=1,beta=1,delta_ij_hat=0.5,gc="normal",
                         delta_gc_list,LL_list,alpha_list,beta_list,save_path="example/output/zibb_fit/",
                         width=10,height=10)

}
